+++ src/hooks/sale/purchaseHistory.ts
@@ -62,7 +62,7 @@ export const usePurchaseHistory = (
                   timestamp: new Date(Number(timestamp)),
                   price: parseInt(price),
                   type: purchaseType,
-                } as PurchaseHistoryItem)
+                }) as PurchaseHistoryItem
             )
           );
         }
